// src/generated/models/Crad2_patientdischargesModel.ts
export const Crad2_patientdischargesand_earlyflag = {
  1: 'Yes',
  0: 'No'
} as const;
export type Crad2_patientdischargesand_earlyflag = keyof typeof Crad2_patientdischargesand_earlyflag;
export const Crad2_patientdischargescrad2_pendingon = {
  408520000: 'Nursing',
  408520001: 'Pharmacy & Billing',
  408520002: 'Front Office',
  408520003: 'Nursing (Physical Discharge)',
  408520004: 'House Keeping'
} as const;
export type Crad2_patientdischargescrad2_pendingon = keyof typeof Crad2_patientdischargescrad2_pendingon;
export const Crad2_patientdischargescrad2_status = {
  408520000: 'Pending',
  408520001: 'In Progress',
  408520002: 'Completed',
  749510001: 'Cancelled',
  749510002: 'Missed'
} as const;
export type Crad2_patientdischargescrad2_status = keyof typeof Crad2_patientdischargescrad2_status;
export const Crad2_patientdischargescrda1_whatsappflag = {
  382720000: '📢'
} as const;
export type Crad2_patientdischargescrda1_whatsappflag = keyof typeof Crad2_patientdischargescrda1_whatsappflag;
export const Crad2_patientdischargesnew_gender = {
  100000000: 'Male',
  100000001: 'Female'
} as const;
export type Crad2_patientdischargesnew_gender = keyof typeof Crad2_patientdischargesnew_gender;
export const Crad2_patientdischargesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crad2_patientdischargesstatecode = keyof typeof Crad2_patientdischargesstatecode;
export const Crad2_patientdischargesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crad2_patientdischargesstatuscode = keyof typeof Crad2_patientdischargesstatuscode;

export interface Crad2_patientdischargesBase {
  and_cancellationreason?: string;
  and_cancelled?: string;
  and_dischargetype?: string;
  and_earlyflag?: Crad2_patientdischargesand_earlyflag;
  and_homemedicationtime?: string;
  and_ishomemedication?: boolean;
  "and_PatientCode@odata.bind"?: string;
  cr301_filedeliverydatetime?: string;
  cr301_patientnationality?: string;
  "cr603_Patient@odata.bind"?: string;
  crad2_admissiondate?: string;
  crad2_bed?: string;
  crad2_dischargedate?: string;
  crad2_dischargereference: string;
  crad2_doctorname?: string;
  crad2_floor?: string;
  crad2_los?: number;
  crad2_patientdischargeid: string;
  crad2_patientidlookup?: string;
  crad2_patientname?: string;
  crad2_pendingon?: Crad2_patientdischargescrad2_pendingon;
  crad2_phonenumber?: string;
  crad2_room?: string;
  crad2_specialty?: string;
  crad2_status?: Crad2_patientdischargescrad2_status;
  crad2_visitid?: string;
  crda1_outpatientclearancedate?: string;
  crda1_whatsappflag?: Crad2_patientdischargescrda1_whatsappflag;
  crda1_whatsappsenddatetime?: string;
  importsequencenumber?: number;
  new_age?: string;
  new_contract?: string;
  new_contractor?: string;
  new_currentstep?: string;
  new_diagnosis?: string;
  new_entityid?: string;
  new_gender?: Crad2_patientdischargesnew_gender;
  new_lastactivity?: string;
  new_location?: string;
  new_nursestation?: string;
  new_outpatientpharmacypending?: string;
  new_patientid?: string;
  new_paymenttype?: string;
  overriddencreatedon?: string;
  statecode: Crad2_patientdischargesstatecode;
  statuscode?: Crad2_patientdischargesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Crad2_patientdischarges extends Crad2_patientdischargesBase {
  and_dischargecondition?: string;
  and_earlyflagname?: string;
  and_homemedicationsdate?: string;
  and_homemedicationsflag?: string;
  and_ishomemedicationname?: string;
  and_patientcodename?: string;
  and_physicaldischargetat?: number;
  and_physicaltat?: string;
  and_physicaltat_date?: string;
  and_physicaltat_state?: number;
  and_physiciandischargedate?: string;
  and_slacategory?: string;
  cr301_filescanstatus?: string;
  cr603_patientname?: string;
  crad2_losdays?: number;
  crad2_pendingonname?: string;
  crad2_statusname?: string;
  crda1_currentstepf?: string;
  crda1_whatsappflagname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  new_gendername?: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  and_patientcode?: object;
  _and_patientcode_value?: string;
  cr603_patient?: object;
  _cr603_patient_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}